restrict denovo kmer size

+++ src/denovo_discovery/denovo_discovery.cpp
@@ -132,7 +132,7 @@ void DenovoDiscovery::find_paths_through_candidate_region(
     remove_graph_file(GATB_graph_filepath);
 }
 
-DenovoDiscovery::DenovoDiscovery(const uint32_t& kmer_size,
+DenovoDiscovery::DenovoDiscovery(const uint16_t& kmer_size,
     const double& read_error_rate, const uint16_t max_insertion_size,
     const uint16_t min_covg_for_node_in_assembly_graph)
     : min_covg_for_node_in_assembly_graph { min_covg_for_node_in_assembly_graph }
